+++ lib/src/log.dart
@@ -1,4 +1,4 @@
 import 'dart:developer' as dev;
-import 'package:kpostal/src/constant.dart';
+import 'package:webview_flutter_kpostal/src/constant.dart';
 
 void log(String msg) => dev.log(msg, name: KpostalConst.packageName);